// src/api/generated/models/RentalResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RentalResponse = {
    listing_id: string;
    full_url: string;
    address: string;
    price: number;
    net_effective_rent: number;
    concession?: (string | null);
    lease_length?: (string | null);
    beds?: (number | null);
    baths?: (number | null);
    listing_agency: string;
    image_url: string;
    longitude?: (number | null);
    latitude?: (number | null);
};